\name{NVC-package}
\alias{NVC-package}
\alias{NVC}
\docType{package}
\title{
\packageTitle{NVC}
}
\description{
\packageDescription{NVC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NVC}
\packageIndices{NVC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NVC}

Maintainer: \packageMaintainer{NVC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
