% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimNVC}
\alias{optimNVC}
\title{Function to optimize the NVCModel}
\usage{
optimNVC(y, mu, U, ellinit)
}
\arguments{
\item{y}{N vector of observations}

\item{mu}{mean vector (N dimensional)}

\item{U}{PQ x Q matrix of P Variance Components}

\item{ellinit}{P vector of initialization values for ell}
}
\value{
List: LogLik is log likelihood at optimima. Pars argument is ells.
}
\description{
Function to optimize the NVCModel
}
\details{
Fits the following model (for unknown ell)
\deqn{y ~ N(mu, e^{ell_1}U_1 + ...+ e^{ell_P}U_1)}
using L-BFGS optimization implemented in optimized C++ code.
}
\examples{
  N <- 1000
  I <- diag(N)
  U1 <- diag(N)
  U2 <-  rWishart(1, N+10, diag(N))[,,1]
  ell <- c(0.0001, 3)
  Sigma <- exp(ell[1])*U1 + exp(ell[2])*U2
  L <- t(chol(Sigma))
  mu <- rep(4, N)
  y <- mu + L \%*\% rnorm(N) 
  fit <- optimNVC(y, mu, rbind(U1, U2), c(0,0))
}
